1. State Pattern 2. GameManager.java, State.java, StateAllEnemiesDead.java, StateAllPlayersDead.java, StateMovingPlayerintoPosition.java, StateOngoingGame.java, StateStart.java, StateMethods.java 3. Because the updateStatus () and perform Status () of the GameManager divide the state by the conditional statement, the state pattern is applied
